+++ src/ionic-tree-provider.ts
@@ -104,8 +104,10 @@ export class IonicTreeProvider implements vscode.TreeDataProvider<Recommendation
       }
     } else {
       const packageJsonPath = path.join(this.workspaceRoot, 'package.json');
-      const folderBased = isFolderBasedMonoRepo(this.workspaceRoot).length > 0;
-      if (this.pathExists(packageJsonPath) || folderBased) {
+      const folders = isFolderBasedMonoRepo(this.workspaceRoot);
+      const folderBased = folders.length > 0;
+      const packageJsonExists = this.pathExists(packageJsonPath);
+      if (packageJsonExists || folderBased) {
         return reviewProject(this.workspaceRoot, this.context, this.selectedProject);
       } else {
         return Promise.resolve(starterProject(this.workspaceRoot));