+++ src/recommend.ts
@@ -52,7 +52,7 @@ export async function getRecommendations(
     const runWeb = new Tip('Web', '', TipType.Run, 'Serve', undefined, 'Running on Web', `Project Served`)
       .setDynamicCommand(ionicServe, project, false)
       .requestIPSelection()
-      .setFeatures([TipFeature.viewInEditor])
+      .setFeatures([TipFeature.viewInEditor, TipFeature.capViewQR])
       .setRunPoints([
         { title: 'Building...', text: 'Generating browser application bundles' },
         { title: 'Serving', text: 'Development server running' },
@@ -306,6 +306,7 @@ export async function getRecommendations(
   }
   project.add(useHttps(project));
   project.add(viewInEditor());
+  project.add(viewAsQR());
 
   // REMOTE LOGGING ENABLED ################
   //project.add(remoteLogging(project));
@@ -396,6 +397,14 @@ function viewInEditor(): Tip {
     .canRefreshAfter();
 }
 
+function viewAsQR(): Tip {
+  const viewAsQR = getSetting(WorkspaceSetting.previewQR);
+  return new Tip('View as QR Code', undefined, viewAsQR ? TipType.Check : TipType.Box, undefined)
+    .setTooltip('Whether the app will be previewed as a QR Code that can be opened with the Cap View App')
+    .setAction(toggleViewQR, viewAsQR)
+    .canRefreshAfter();
+}
+
 function toggleRemoteLogging(project: Project, current: boolean): Promise<void> {
   if (startLogServer(project.folder)) {
     ionicState.remoteLogging = !current;
@@ -425,4 +434,8 @@ async function toggleHttps(current: boolean, project: Project) {
 
 async function toggleViewInEditor(current: boolean) {
   await setSetting(WorkspaceSetting.previewInEditor, !current);
+}
+
+async function toggleViewQR(current: boolean) {
+  await setSetting(WorkspaceSetting.previewQR, !current);
 }